♻️ refactor:  AppData周りの整理
Issue #285
