+++ lib/domain/repository/app_data/app_data_repository.dart
@@ -4,31 +4,11 @@
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../repo/entity/search_repos_order.dart';
-import '../repo/entity/search_repos_sort.dart';
 import 'entity/app_data_key.dart';
 
 /// アプリデータRepositoryプロバイダー
 final appDataRepositoryProvider = Provider<AppDataRepository>(
   (ref) => throw UnimplementedError('Provider was not initialized'),
-);
-
-/// リポジトリ検索用オーダー値プロバイダー
-final searchReposOrderProvider =
-    StateNotifierProvider<SearchReposOrderNotifier, SearchReposOrder>(
-  (ref) => SearchReposOrderNotifier(
-    appDataRepository: ref.watch(appDataRepositoryProvider),
-  ),
-  name: 'searchReposOrderProvider',
-);
-
-/// リポジトリ検索用ソート値プロバイダー
-final searchReposSortProvider =
-    StateNotifierProvider<SearchReposSortNotifier, SearchReposSort>(
-  (ref) => SearchReposSortNotifier(
-    appDataRepository: ref.watch(appDataRepositoryProvider),
-  ),
-  name: 'searchReposSortProvider',
 );
 
 /// アプリデータRepository
@@ -41,43 +21,4 @@ abstract class AppDataRepository {
 
   /// Stream を返す
   Stream<T> changes<T extends Object?>(AppDataKey key);
-}
-
-/// アプリデータNotifier
-abstract class AppDataNotifier<T> extends StateNotifier<T> {
-  AppDataNotifier({
-    required this.appDataRepository,
-    required this.appDataKey,
-  }) : super(
-          appDataRepository.get<T>(appDataKey),
-        ) {
-    // Streamを監視して都度反映する
-    appDataRepository.changes<T>(appDataKey).listen(
-      (value) {
-        state = value;
-      },
-    );
-  }
-
-  final AppDataRepository appDataRepository;
-  final AppDataKey appDataKey;
-
-  /// 設定する
-  void set(T value) {
-    appDataRepository.set<T>(appDataKey, value);
-  }
-}
-
-/// リポジトリ検索用オーダー値Notifier
-class SearchReposOrderNotifier extends AppDataNotifier<SearchReposOrder> {
-  SearchReposOrderNotifier({
-    required super.appDataRepository,
-  }) : super(appDataKey: AppDataKey.searchReposOrder);
-}
-
-/// リポジトリ検索用ソート値Notifier
-class SearchReposSortNotifier extends AppDataNotifier<SearchReposSort> {
-  SearchReposSortNotifier({
-    required super.appDataRepository,
-  }) : super(appDataKey: AppDataKey.searchReposSort);
 }
